+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `sprocket --no-ignore` disables `.sprocketignore` processing while WDL
   documents are discovered ([#1110](https://github.com/stjude-rust-labs/sprocket/pull/1110)).
 * `sprocket dev test` now produces spanned diagnostics for YAML files ([#982](https://github.com/stjude-rust-labs/sprocket/pull/982)).
+* `run` and `dev test` now show status bars for container image pulls ([#1117](https://github.com/stjude-rust-labs/sprocket/pull/1117)).
 
 ### Changed