+++ crates/wdl-engine/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+<!-- TODO: Add PR number -->
+* Added support for the `ImagePull{Started, Failed, Finished}` `crankshaft` events ([#]()).
+
 #### Changed
 
 * `EngineEvent` now emits `TaskInitializing` before an execution attempt and